Refactored SkyMapAnnotator in a way that Framing Entities know how to draw themselves, to move there code out of the annotator.
Former-commit-id: c4902b4e3fcbc298a3f77882a4a4c53c4db2e682
Former-commit-id: 6b989652b15d2cf40292cc508605d978186e4b18

+++ NINA/Utility/DatabaseInteraction.cs
@@ -176,7 +176,7 @@ namespace NINA.Utility {
                         while (reader.Read()) {
                             starList.Add(new Star(reader.GetInt32(0), reader.GetString(1),
                                 new Coordinates(reader.GetDouble(2), reader.GetDouble(3), Epoch.J2000,
-                                    Coordinates.RAType.Degrees), reader.GetDouble(4)));
+                                    Coordinates.RAType.Degrees), (float)reader.GetDouble(4)));
                         }
                     }
